pyproject.toml
[project]
name = "ai-climate-overview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-tavily>=0.2.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
